+++ controllers/userCtrl.js
@@ -1,0 +1,253 @@
+import User from "../models/userModel.js";
+import asyncHandler from "express-async-handler";
+import generateToken from "../util/generateToken.js";
+
+const signUp = asyncHandler(async (req, res) => {
+  const { email, mName, password, type } = req.body;
+  try {
+    const estimate = await User.estimatedDocumentCount();
+    if (estimate > 0) {
+      const existingUser = await User.findOne({ email });
+      if (existingUser) {
+        return res.json({
+          success: false,
+          message: "User with this email already exists",
+        });
+      }
+      const newUser = new User({ email, mName, password, type });
+      await newUser.save();
+      res.json({
+        success: true,
+        message: "Account created successfully",
+        userId: newUser._id,
+      });
+    } else {
+      const newUser = new User({ email, mName, password, type });
+      await newUser.save();
+      // const newAdmin = new Admin({ email, mName, type: 'main' });
+      // await newAdmin.save();
+      res.json({
+        success: true,
+        message: "Account created successfully",
+        userId: newUser._id,
+      });
+    }
+  } catch (error) {
+    console.log(error);
+  }
+});
+
+const googleAuth = asyncHandler(async (req, res) => {
+  const { name, email, token } = req.body;
+  try {
+    const existingUser = await User.findOne({ email });
+    if (existingUser) {
+      if (!existingUser.password) {
+        return res.json({
+          success: true,
+          message: "Google authentication successful",
+          userData: existingUser,
+          passwordSetRequired: true,
+        });
+      }
+      return res.json({
+        success: true,
+        message: "SignIn successful",
+        userData: existingUser,
+      });
+    }
+    const newUser = new User({ email, mName: name, resetPasswordToken: token });
+    await newUser.save();
+    res.json({
+      success: true,
+      message: "Account created successfully Please set Password",
+      userData: newUser,
+    });
+  } catch (error) {
+    console.log(error);
+  }
+});
+
+const signIn = asyncHandler(async (req, res) => {
+  const { email, password } = req.body;
+
+  try {
+    const user = await User.findOne({ email });
+
+    if (!user) {
+      return res.json({ success: false, message: "Invalid email or password" });
+    }
+
+    if (password === user.password) {
+      generateToken(res, user._id);
+      return res.json({
+        success: true,
+        message: "SignIn successful",
+        userData: user,
+      });
+    }
+
+    res.json({ success: false, message: "Invalid email or password" });
+  } catch (error) {
+    console.log(error);
+  }
+});
+const userDetails = asyncHandler(async (req, res) => {
+  try {
+    // Fetch the admin from the database (for simplicity, we'll assume there's only one admin)
+    const admin = await User.findOne(req.body); // You can modify this to look up specific admins
+
+    // If no admin is found, return a 404 error
+    if (!admin) {
+      return res.status(404).json({ message: "Admin not found" });
+    }
+
+    // Send the admin data back to the frontend
+    res.json({ admin });
+  } catch (error) {
+    // Handle errors and send a 500 status
+    console.log(error);
+  }
+});
+
+const forgetPassword = asyncHandler(async (req, res) => {
+  const { email, name, company, logo } = req.body;
+
+  try {
+    const user = await User.findOne({ email });
+
+    if (!user) {
+      return res.json({ success: false, message: "User not found" });
+    }
+
+    const token = crypto.randomBytes(20).toString("hex");
+    user.resetPasswordToken = token;
+    user.resetPasswordExpires = Date.now() + 3600000;
+    await user.save();
+
+    const resetLink = `${process.env.WEBSITE_URL}/reset-password/${token}`;
+
+    const mailOptions = {
+      from: process.env.EMAIL,
+      to: user.email,
+      subject: `${name} Password Reset`,
+      html: `<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+              <meta http-equiv="Content-Type" content="text/html charset=UTF-8" />
+              <html lang="en">
+              
+                <head></head>
+               <div id="__react-email-preview" style="display:none;overflow:hidden;line-height:1px;opacity:0;max-height:0;max-width:0">Password Reset<div> ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿ ‌​‍‎‏﻿</div>
+               </div>
+              
+                <body style="margin-left:auto;margin-right:auto;margin-top:auto;margin-bottom:auto;background-color:rgb(255,255,255);font-family:ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont, &quot;Segoe UI&quot;, Roboto, &quot;Helvetica Neue&quot;, Arial, &quot;Noto Sans&quot;, sans-serif, &quot;Apple Color Emoji&quot;, &quot;Segoe UI Emoji&quot;, &quot;Segoe UI Symbol&quot;, &quot;Noto Color Emoji&quot;">
+                  <table align="center" role="presentation" cellSpacing="0" cellPadding="0" border="0" width="100%" style="max-width:37.5em;margin-left:auto;margin-right:auto;margin-top:40px;margin-bottom:40px;width:465px;border-radius:0.25rem;border-width:1px;border-style:solid;border-color:rgb(234,234,234);padding:20px">
+                    <tr style="width:100%">
+                      <td>
+                        <table align="center" border="0" cellPadding="0" cellSpacing="0" role="presentation" width="100%" style="margin-top:32px">
+                          <tbody>
+                            <tr>
+                              <td><img alt="Vercel" src="${logo}" width="40" height="37" style="display:block;outline:none;border:none;text-decoration:none;margin-left:auto;margin-right:auto;margin-top:0px;margin-bottom:0px" /></td>
+                            </tr>
+                          </tbody>
+                        </table>
+                        <h1 style="margin-left:0px;margin-right:0px;margin-top:30px;margin-bottom:30px;padding:0px;text-align:center;font-size:24px;font-weight:400;color:rgb(0,0,0)">Password Reset</h1>
+                        <p style="font-size:14px;line-height:24px;margin:16px 0;color:rgb(0,0,0)">Click on the button below to reset the password for your account ${email}.</p>
+                        <table align="center" border="0" cellPadding="0" cellSpacing="0" role="presentation" width="100%" style="margin-bottom:32px;margin-top:32px;text-align:center">
+                          <tbody>
+                            <tr>
+                              <td><a href="${resetLink}" target="_blank" style="p-x:20px;p-y:12px;line-height:100%;text-decoration:none;display:inline-block;max-width:100%;padding:12px 20px;border-radius:0.25rem;background-color:rgb(0,0,0);text-align:center;font-size:12px;font-weight:600;color:rgb(255,255,255);text-decoration-line:none"><span></span><span style="p-x:20px;p-y:12px;max-width:100%;display:inline-block;line-height:120%;text-decoration:none;text-transform:none;mso-padding-alt:0px;mso-text-raise:9px"</span><span>Reset</span></a></td>
+                            </tr>
+                          </tbody>
+                        </table>
+                        <p style="font-size:14px;line-height:24px;margin:16px 0;color:rgb(0,0,0)">Best,<p target="_blank" style="color:rgb(0,0,0);text-decoration:none;text-decoration-line:none">The <strong>${company}</strong> Team</p></p>
+                        </td>
+                    </tr>
+                  </table>
+                </body>
+              
+              </html>`,
+    };
+
+    await transporter.sendMail(mailOptions);
+
+    res.json({
+      success: true,
+      message: "Password reset link sent to your email",
+    });
+  } catch (error) {
+    res.status(500).json({ success: false, message: "Internal server error" });
+  }
+});
+
+const userProfile = asyncHandler(async (req, res) => {
+  const { email, mName, password, uid } = req.body;
+  try {
+    if (password === "") {
+      await User.findOneAndUpdate(
+        { _id: uid },
+        { $set: { email: email, mName: mName } }
+      )
+        .then((result) => {
+          res.json({ success: true, message: "Profile Updated" });
+        })
+        .catch((error) => {
+          res
+            .status(500)
+            .json({ success: false, message: "Internal server error" });
+        });
+    } else {
+      await User.findOneAndUpdate(
+        { _id: uid },
+        { $set: { email: email, mName: mName, password: password } }
+      )
+        .then((result) => {
+          res.json({ success: true, message: "Profile Updated" });
+        })
+        .catch((error) => {
+          res
+            .status(500)
+            .json({ success: false, message: "Internal server error" });
+        });
+    }
+  } catch (error) {
+    res.status(500).json({ success: false, message: "Internal server error" });
+  }
+});
+
+const resetPassword = asyncHandler(async (req, res) => {
+  const { password, token } = req.body;
+
+  try {
+    const user = await User.findOne({
+      resetPasswordToken: token,
+      resetPasswordExpires: { $gt: Date.now() },
+    });
+
+    if (!user) {
+      return res.json({ success: true, message: "Invalid or expired token" });
+    }
+
+    user.password = password;
+    user.resetPasswordToken = null;
+    user.resetPasswordExpires = null;
+
+    await user.save();
+
+    res.json({
+      success: true,
+      message: "Password updated successfully",
+      email: user.email,
+    });
+  } catch (error) {
+    res.status(500).json({ success: false, message: "Internal server error" });
+  }
+});
+export {
+  signUp,
+  googleAuth,
+  signIn,
+  userDetails,
+  forgetPassword,
+  resetPassword,
+  userProfile,
+};
