+++ package.json
@@ -2,6 +2,7 @@
   "name": "ai-based-training-platform1-backend",
   "version": "1.0.0",
   "main": "server.js",
+  "type": "module",
   "scripts": {
     "server": "nodemon server.js",
     "deploy": "firebase deploy --only functions:api"
@@ -18,9 +19,11 @@
     "crypto": "^1.0.1",
     "dotenv": "^16.4.5",
     "express": "^4.21.0",
+    "express-async-handler": "^1.2.0",
     "firebase-admin": "^12.5.0",
     "firebase-functions": "^6.0.1",
     "g-i-s": "^2.1.7",
+    "jsonwebtoken": "^9.0.2",
     "mongoose": "^8.7.0",
     "nodemailer": "^6.9.15",
     "showdown": "^2.1.0",
